+++ src/components/Heading/Heading.tsx
@@ -10,6 +10,8 @@ import {
 import { HeadingDisplaySizes, HeadingLevels } from "./HeadingTypes";
 
 export interface HeadingProps {
+  /** Optionally pass in additional Chakra-based styles. */
+  additionalStyles?: { [key: string]: any };
   /** Optional className that appears in addition to `heading` */
   className?: string;
   /** Optional size used to override the default styles of the semantic HTML `<h>` elements */
@@ -18,8 +20,6 @@ export interface HeadingProps {
   id?: string;
   /** Optional number 1-6 used to create the `<h*>` tag; if prop is not passed, `Heading` will default to `<h2>` */
   level?: HeadingLevels;
-  /** Optionally pass in additional Chakra-based styles. */
-  sx?: { [key: string]: any };
   /** Inner text of the `<h*>` element */
   text?: string;
   /** Optional URL that header points to; when `url` prop is passed to `Heading`, a child `<a>` element is created and the heading text becomes an active link */
@@ -43,11 +43,11 @@ const getVariant = (displaySize) => variantMap[displaySize] || null;
 
 function Heading(props: React.PropsWithChildren<HeadingProps>) {
   const {
+    additionalStyles = {},
     className,
     displaySize,
     id,
     level = HeadingLevels.Two,
-    sx = {},
     text,
     url,
     urlClass,
@@ -55,8 +55,8 @@ function Heading(props: React.PropsWithChildren<HeadingProps>) {
   const variant = displaySize ? getVariant(displaySize) : `h${level}`;
   const styles = useStyleConfig("Heading", { variant });
   // Combine native base styles with any additional styles.
-  // This is used only in the `Hero` component, for now.
-  const finalStyles = { ...styles, ...sx };
+  // This is used in the `Hero` and `Notification` components.
+  const finalStyles = { ...styles, ...additionalStyles };
   const asHeading: any = `h${level}`;
 
   if (level < 1 || level > 6) {